requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
progressbar2==3.51.0
pycurl==7.43.0.5
python-utils==2.4.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
